fix: Renaming to CID and renaming file. #4
