+++ lib/src/decode_cid.dart
